refactor: Move drng/info api to drng/
